move inconsistencies_between_fvar_stat out of fontwerk directory
